Fix selector menu bug
