+++ snappet-challenge/src/assets/subjectsMockUp.js
@@ -1,0 +1,71 @@
+const subjectsMockUp = [
+  {
+    id: 1,
+    title: "Math Grade 5",
+    learningObjectives: [
+      {
+        id: 1,
+        title: "001. Learning Objective number one",
+        pupils: {
+          present: 16,
+          absent: 0,
+          total: 40,
+        },
+      },
+      {
+        id: 2,
+        title: "002. Learning Objective number two",
+        pupils: {
+          present: 3,
+          absent: 3,
+          total: 40,
+        },
+      },
+      {
+        id: 3,
+        title: "003. Learning Objective number three",
+        pupils: {
+          present: 3,
+          absent: 3,
+          total: 40,
+        },
+      },
+    ],
+  },
+
+  {
+    id: 2,
+    title: "Math Grade 4",
+    learningObjectives: [
+      {
+        id: 1,
+        title: "004. Learning Objective number four",
+        pupils: {
+          present: 10,
+          absent: 6,
+          total: 40,
+        },
+      },
+      {
+        id: 2,
+        title: "005. Learning Objective number five",
+        pupils: {
+          present: 5,
+          absent: 1,
+          total: 40,
+        },
+      },
+      {
+        id: 3,
+        title: "006. Learning Objective number six",
+        pupils: {
+          present: 20,
+          absent: 4,
+          total: 40,
+        },
+      },
+    ],
+  },
+];
+
+export default subjectsMockUp;
